Render all stats in one page.
After a bit pondering about the use cases, this is absolutely the best
solution. If there's too much detailed info with expenses, those can be
hidden dynamically.

Also remove tests as most of them are covered by the integration test.
Will add more meat in the future so we should end up with a pretty good
tests coverage.