Simplify GetSalaryCompensatedDebts function.
When salary compensated debt is needed and order must be kept in the
original data structure, this is achieved with pointer parameters. By
utilizing the slice inside the function we can order the items as we
like and modify then as needed since they are passed as references.
Otherwise code would become a dirty if-else jungle.