Simplify debt control.
Now we leverage a pointer parameters and external data structures in
debt control to get debts updated. It was previously a tad off because
my memory served me wrong. Values used to be in certain order but I've
changed that lately. Should have checked earlier.

Will fix the tests later.